Sithafal Task - 6/requirements.txt
llama-cloud
pydantic
python-dotenv
